Adding persistance support using the configured H2 database - Updated controllers and services to work with persistance changes. - Added a sequence generator to generate unique ids, this is done so that we can do only one database operation   when saving the long url and tiny expression, this was inspired from a blog post linked. - Added JPA entity and repository for Link. - Updated unit and integration tests to work with database changes. - Updated README file with instructions to access h2 console.

+++ src/main/java/com/backbase/tinyexpression/util/IDConverter.java
@@ -10,14 +10,14 @@ public class IDConverter {
      * @param id the id.
      * @return the tiny expression.
      */
-    public static String idToTinyExpression(int id) {
+    public static String idToTinyExpression(long id) {
 
         char map[] = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789".toCharArray();
 
         StringBuilder shortURL = new StringBuilder();
 
         while(id > 0) {
-            shortURL.append(map[id % 62]);
+            shortURL.append(map[(int) (id % 62)]);
             id = id / 62;
         }
         return shortURL.reverse().toString();
@@ -25,11 +25,11 @@ public class IDConverter {
 
     /**
      * Helper method to get the id using the tiny expression.
-     * @param tinyExpression
-     * @return
+     * @param tinyExpression tiny expression.
+     * @return the id.
      */
-    public static int tinyExpressionToID(String tinyExpression) {
-        int id = 0;
+    public static long tinyExpressionToID(String tinyExpression) {
+        long id = 0;
 
         for(char ch : tinyExpression.toCharArray()) {
             if('a' <= ch && ch <= 'z'){
